requirements.txt
asgiref==3.3.4
Django==3.2
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.3.0
